Updated crate modules visibility
