+++ main/src/adapters.rs
