--- main/src/preprocessing/tokenizer/mod.rs
@@ -1,25 +1,0 @@
-// Copyright 2019 Guillaume Becquin
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//     http://www.apache.org/licenses/LICENSE-2.0
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-pub mod albert_tokenizer;
-pub mod base_tokenizer;
-pub mod bert_tokenizer;
-mod constants;
-pub mod ctrl_tokenizer;
-pub mod gpt2_tokenizer;
-pub mod marian_tokenizer;
-pub mod openai_gpt_tokenizer;
-pub mod roberta_tokenizer;
-pub mod sentence_piece_tokenizer;
-pub mod t5_tokenizer;
-pub mod tokenization_utils;
-pub mod xlm_roberta_tokenizer;
-pub mod xlnet_tokenizer;
