--- main/src/preprocessing/vocab/mod.rs
@@ -1,24 +1,0 @@
-// Copyright 2019 Guillaume Becquin
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//     http://www.apache.org/licenses/LICENSE-2.0
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-pub mod albert_vocab;
-pub mod base_vocab;
-pub mod bert_vocab;
-pub mod bpe_vocab;
-pub mod gpt2_vocab;
-pub mod marian_vocab;
-pub mod openai_gpt_vocab;
-pub mod roberta_vocab;
-pub mod sentence_piece_vocab;
-pub mod sentencepiece_proto;
-pub mod t5_vocab;
-pub mod xlm_roberta_vocab;
-pub mod xlnet_vocab;
