+++ main/src/tokenizer/openai_gpt_tokenizer.rs
@@ -12,165 +12,129 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::preprocessing::error::TokenizerError;
-use crate::preprocessing::tokenizer::base_tokenizer::{Mask, Token, TokenRef, Tokenizer};
-use crate::preprocessing::tokenizer::constants::UNICODE_TO_BYTES;
-use crate::preprocessing::tokenizer::tokenization_utils::{
-    bpe, fix_mask, split_on_bpe_pairs, split_on_regex_with_lookahead, split_on_special_tokens,
-};
-use crate::preprocessing::vocab::base_vocab::Vocab;
-use crate::preprocessing::vocab::bpe_vocab::BpePairVocab;
-use crate::tokenization_utils::lowercase;
-use crate::Gpt2Vocab;
-use itertools::Itertools;
-use regex::Regex;
+use crate::error::TokenizerError;
+use crate::tokenizer::tokenization_utils::{openai_gpt_bpe, split_on_bpe_pairs};
+use crate::tokenizer::{BaseTokenizer, Tokenizer};
+use crate::vocab::bpe_vocab::BpePairVocab;
+use crate::vocab::{OpenAiGptVocab, Vocab};
+use crate::{Mask, Token, TokenRef};
 use std::cell::RefCell;
 use std::collections::HashMap;
-use std::iter::Iterator;
 use std::rc::Rc;
+use std::sync::Arc;
 
-pub struct Gpt2Tokenizer {
-    vocab: Rc<Gpt2Vocab>,
+pub struct OpenAiGptTokenizer {
+    vocab: Arc<OpenAiGptVocab>,
+    base_tokenizer: BaseTokenizer<OpenAiGptVocab>,
     bpe_ranks: Rc<BpePairVocab>,
     cache: RefCell<HashMap<String, (Vec<String>, Vec<usize>)>>,
-    pattern_lookahead: Regex,
-    pattern_tokenization: Regex,
-    lower_case: bool,
 }
 
-impl Gpt2Tokenizer {
+impl OpenAiGptTokenizer {
     pub fn from_file(
         vocab_path: &str,
         merges_path: &str,
         lower_case: bool,
-    ) -> Result<Gpt2Tokenizer, TokenizerError> {
-        let vocab = Rc::new(Gpt2Vocab::from_file(vocab_path)?);
+    ) -> Result<OpenAiGptTokenizer, TokenizerError> {
+        let vocab = Arc::new(OpenAiGptVocab::from_file(vocab_path)?);
+        let base_tokenizer = BaseTokenizer::from_existing_vocab(vocab.clone(), lower_case, true);
         let bpe_ranks = Rc::new(BpePairVocab::from_file(merges_path)?);
         let cache = RefCell::new(HashMap::new());
-        let pattern_lookahead = Regex::new(r"\s+\S").unwrap();
-        let pattern_tokenization =
-            Regex::new(r"'s|'t|'re|'ve|'m|'ll|'d| ?\p{L}+| ?\p{N}+| ?[^\s\p{L}\p{N}]+|\s+")
-                .unwrap();
-        Ok(Gpt2Tokenizer {
+        Ok(OpenAiGptTokenizer {
             vocab,
+            base_tokenizer,
             bpe_ranks,
             cache,
-            pattern_lookahead,
-            pattern_tokenization,
-            lower_case,
         })
     }
 
     pub fn from_existing_vocab_and_merges(
-        vocab: Rc<Gpt2Vocab>,
+        vocab: Arc<OpenAiGptVocab>,
         merges: Rc<BpePairVocab>,
         lower_case: bool,
-    ) -> Gpt2Tokenizer {
+    ) -> OpenAiGptTokenizer {
+        let base_tokenizer = BaseTokenizer::from_existing_vocab(vocab.clone(), lower_case, true);
         let cache = RefCell::new(HashMap::new());
-        let pattern_lookahead = Regex::new(r"\s+\S").unwrap();
-        let pattern_tokenization =
-            Regex::new(r"'s|'t|'re|'ve|'m|'ll|'d| ?\p{L}+| ?\p{N}+| ?[^\s\p{L}\p{N}]+|\s+")
-                .unwrap();
-        Gpt2Tokenizer {
+        OpenAiGptTokenizer {
             vocab,
+            base_tokenizer,
             bpe_ranks: merges,
             cache,
-            pattern_lookahead,
-            pattern_tokenization,
-            lower_case,
         }
     }
 }
 
-impl Tokenizer<Gpt2Vocab> for Gpt2Tokenizer {
-    fn vocab(&self) -> &Gpt2Vocab {
+impl Tokenizer<OpenAiGptVocab> for OpenAiGptTokenizer {
+    fn vocab(&self) -> &OpenAiGptVocab {
         self.vocab.as_ref()
     }
 
     fn tokenize_to_tokens(&self, initial_token: TokenRef) -> Vec<Token> {
-        let mut tokens = split_on_special_tokens(initial_token, self.vocab.as_ref())
+        let tokens: Vec<Token> = self
+            .base_tokenizer
+            .tokenize_to_tokens(initial_token)
             .into_iter()
-            .map(|token| token.to_owned())
-            .collect::<Vec<Token>>();
-
-        let mut sub_tokens = Vec::new();
-        for token in tokens.iter_mut() {
-            if token.mask != Mask::Special && token.mask != Mask::Unknown {
-                if self.lower_case {
-                    lowercase(token);
-                }
-                for token in split_on_regex_with_lookahead(
-                    token.as_ref(),
-                    &self.pattern_lookahead,
-                    &self.pattern_tokenization,
-                ) {
-                    sub_tokens.extend(split_on_bpe_pairs(
-                        token,
-                        bpe,
+            .map(|token| {
+                if token.mask != Mask::Special && token.mask != Mask::Unknown {
+                    split_on_bpe_pairs(
+                        token.as_ref(),
+                        openai_gpt_bpe,
                         (&self.bpe_ranks).as_ref(),
                         &self.cache,
-                        true,
-                    ));
+                        false,
+                    )
+                } else {
+                    vec![token]
                 }
-            } else {
-                sub_tokens.push(token.clone());
-            }
-        }
+            })
+            .flatten()
+            .collect();
 
-        fix_mask(&mut sub_tokens);
-        sub_tokens
+        tokens
     }
 
     fn convert_tokens_to_string(&self, tokens: Vec<String>) -> String {
-        let tokens = tokens
-            .iter()
-            .join("")
-            .replace(" ##", "")
-            .trim()
-            .chars()
-            .map(|character| *UNICODE_TO_BYTES.get(&character).unwrap())
-            .collect::<Vec<u8>>();
-        String::from_utf8_lossy(tokens.as_slice()).to_string()
+        tokens.join("").replace("</w>", " ").trim().to_owned()
     }
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::preprocessing::tokenizer::base_tokenizer::{
-        Offset, TokenizedInput, TruncationStrategy,
-    };
-    use crate::preprocessing::vocab::base_vocab::swap_key_values;
-    use crate::Gpt2Vocab;
+    use crate::tokenizer::base_tokenizer::{Offset, TokenizedInput, TruncationStrategy};
+    use crate::vocab::base_vocab::swap_key_values;
+    use crate::vocab::OpenAiGptVocab;
+    use itertools::Itertools;
     use std::collections::HashMap;
 
-    fn generate_test_vocab() -> Gpt2Vocab {
+    fn generate_test_vocab() -> OpenAiGptVocab {
         let values: HashMap<String, i64> = [
             ("t".to_owned(), 0),
             ("h".to_owned(), 1),
-            ("a@@".to_owned(), 2),
+            ("a</w>".to_owned(), 2),
             ("n".to_owned(), 3),
             ("the".to_owned(), 4),
             ("Ġ".to_owned(), 5),
-            ("<|endoftext|>".to_owned(), 6),
-            ("o@@".to_owned(), 7),
-            ("Ġear".to_owned(), 8),
-            ("th".to_owned(), 9),
+            ("<unk>".to_owned(), 6),
+            ("o</w>".to_owned(), 7),
+            ("the</w>".to_owned(), 8),
+            ("rth</w>".to_owned(), 9),
+            ("ea".to_owned(), 10),
         ]
         .iter()
         .cloned()
         .collect();
 
         let special_values: HashMap<String, i64> =
-            [("<|endoftext|>".to_owned(), 6)].iter().cloned().collect();
+            [("<unk>".to_owned(), 6)].iter().cloned().collect();
 
         let indices = swap_key_values(&values);
         let special_indices = swap_key_values(&special_values);
 
-        Gpt2Vocab {
+        OpenAiGptVocab {
             values,
             indices,
-            unknown_value: "<|endoftext|>",
+            unknown_value: "<unk>",
             special_values,
             special_indices,
         }
@@ -178,16 +142,15 @@ mod tests {
 
     fn generate_test_merges() -> BpePairVocab {
         let values: HashMap<(String, String), i64> = [
-            (("Ġ".to_owned(), "t".to_owned()), 0),
-            (("Ġ".to_owned(), "n".to_owned()), 1),
-            (("e".to_owned(), "e".to_owned()), 2),
-            (("Ġt".to_owned(), "he".to_owned()), 3),
+            (("4".to_owned(), "t".to_owned()), 0),
+            (("2".to_owned(), "n".to_owned()), 1),
+            (("r".to_owned(), "th</w>".to_owned()), 2),
+            (("t".to_owned(), "he</w>".to_owned()), 3),
             (("h".to_owned(), "e".to_owned()), 4),
-            (("t".to_owned(), "h".to_owned()), 5),
-            (("t".to_owned(), "he".to_owned()), 6),
-            (("Ġ".to_owned(), "e".to_owned()), 7),
-            (("Ġe".to_owned(), "a".to_owned()), 8),
-            (("Ġea".to_owned(), "r".to_owned()), 9),
+            (("t".to_owned(), "h</w>".to_owned()), 5),
+            (("t".to_owned(), "h".to_owned()), 6),
+            (("th".to_owned(), "e</w>".to_owned()), 7),
+            (("e".to_owned(), "a".to_owned()), 8),
         ]
         .iter()
         .cloned()
@@ -197,18 +160,16 @@ mod tests {
     }
 
     #[test]
-    fn test_gpt2_tokenizer() {
+    fn test_openai_gpt_tokenizer() {
         //        Given
-        let vocab = Rc::new(generate_test_vocab());
+        let vocab = Arc::new(generate_test_vocab());
         let merges = Rc::new(generate_test_merges());
-        let gpt2_tokenizer: Gpt2Tokenizer =
-            Gpt2Tokenizer::from_existing_vocab_and_merges(vocab, merges, true);
+        let openai_gpt_tokenizer: OpenAiGptTokenizer =
+            OpenAiGptTokenizer::from_existing_vocab_and_merges(vocab, merges, true);
         let test_tuples = [
-            ("the Earth", vec!["the", "Ġear", "th"]),
+            ("The earth", vec!["the</w>", "ea", "rth</w>"]),
             ("", vec![]),
             (" ", vec![]),
-            ("   t", vec!["Ġ", "Ġ", "Ġt"]),
-            ("t ", vec!["t", "Ġ"]),
             (" \n ", vec![]),
         ];
         let source_texts: Vec<&str> = test_tuples.iter().map(|v| v.0).collect();
@@ -216,27 +177,29 @@ mod tests {
 
         //        When & Then
         for (source_text, expected_result) in test_tuples.iter() {
-            assert_eq!(gpt2_tokenizer.tokenize(*source_text), *expected_result);
+            assert_eq!(
+                openai_gpt_tokenizer.tokenize(*source_text),
+                *expected_result
+            );
         }
 
         assert_eq!(
-            gpt2_tokenizer.tokenize_list(source_texts.clone()),
+            openai_gpt_tokenizer.tokenize_list(source_texts.clone()),
             expected_results
         );
     }
 
     #[test]
-    fn test_gpt2_tokenizer_no_lower_casing() {
+    fn test_openai_gpt_tokenizer_no_lower_casing() {
         //        Given
-        let vocab = Rc::new(generate_test_vocab());
+        let vocab = Arc::new(generate_test_vocab());
         let merges = Rc::new(generate_test_merges());
-        let gpt2_tokenizer: Gpt2Tokenizer =
-            Gpt2Tokenizer::from_existing_vocab_and_merges(vocab, merges, false);
+        let openai_gpt_tokenizer: OpenAiGptTokenizer =
+            OpenAiGptTokenizer::from_existing_vocab_and_merges(vocab, merges, false);
         let test_tuples = [
-            ("the Earth", vec!["the", "Ġ", "E", "a", "r", "th"]),
+            ("The Earth", vec!["T", "h", "e</w>", "E", "a", "rth</w>"]),
             ("", vec![]),
             (" ", vec![]),
-            ("   t", vec!["Ġ", "Ġ", "Ġt"]),
             (" \n ", vec![]),
         ];
         let source_texts: Vec<&str> = test_tuples.iter().map(|v| v.0).collect();
@@ -244,11 +207,14 @@ mod tests {
 
         //        When & Then
         for (source_text, expected_result) in test_tuples.iter() {
-            assert_eq!(gpt2_tokenizer.tokenize(*source_text), *expected_result);
+            assert_eq!(
+                openai_gpt_tokenizer.tokenize(*source_text),
+                *expected_result
+            );
         }
 
         assert_eq!(
-            gpt2_tokenizer.tokenize_list(source_texts.clone()),
+            openai_gpt_tokenizer.tokenize_list(source_texts.clone()),
             expected_results
         );
     }
@@ -256,26 +222,26 @@ mod tests {
     #[test]
     fn test_encode() {
         //        Given
-        let vocab = Rc::new(generate_test_vocab());
+        let vocab = Arc::new(generate_test_vocab());
         let merges = Rc::new(generate_test_merges());
-        let gpt2_tokenizer: Gpt2Tokenizer =
-            Gpt2Tokenizer::from_existing_vocab_and_merges(vocab, merges, true);
+        let openai_gpt_tokenizer: OpenAiGptTokenizer =
+            OpenAiGptTokenizer::from_existing_vocab_and_merges(vocab, merges, true);
         let truncation_strategy = TruncationStrategy::LongestFirst;
         let test_tuples = [
             (
                 "the earth",
                 TokenizedInput {
-                    token_ids: vec![4, 8, 9],
+                    token_ids: vec![8, 10, 9],
                     segment_ids: vec![0, 0, 0],
                     special_tokens_mask: vec![0, 0, 0],
                     overflowing_tokens: vec![],
                     num_truncated_tokens: 0,
                     token_offsets: vec![
                         Some(Offset { begin: 0, end: 3 }),
-                        Some(Offset { begin: 3, end: 7 }),
-                        Some(Offset { begin: 7, end: 9 }),
+                        Some(Offset { begin: 4, end: 6 }),
+                        Some(Offset { begin: 6, end: 9 }),
                     ],
-                    reference_offsets: vec![vec![0, 1, 2], vec![3, 4, 5, 6], vec![7, 8]],
+                    reference_offsets: vec![vec![0, 1, 2], vec![4, 5], vec![6, 7, 8]],
                     mask: vec![Mask::None, Mask::Begin, Mask::Continuation],
                 },
             ),
@@ -313,12 +279,12 @@ mod tests {
         //        When & Then
         for (source_text, expected_result) in test_tuples.iter() {
             assert_eq!(
-                gpt2_tokenizer.encode(source_text, None, 128, &truncation_strategy, 0),
+                openai_gpt_tokenizer.encode(source_text, None, 128, &truncation_strategy, 0),
                 *expected_result
             );
         }
         assert_eq!(
-            gpt2_tokenizer.encode_list(source_texts.clone(), 128, &truncation_strategy, 0),
+            openai_gpt_tokenizer.encode_list(source_texts.clone(), 128, &truncation_strategy, 0),
             expected_results
         );
     }
@@ -326,20 +292,20 @@ mod tests {
     #[test]
     fn test_decode() {
         //        Given
-        let vocab = Rc::new(generate_test_vocab());
+        let vocab = Arc::new(generate_test_vocab());
         let merges = Rc::new(generate_test_merges());
-        let gpt2_tokenizer: Gpt2Tokenizer =
-            Gpt2Tokenizer::from_existing_vocab_and_merges(vocab, merges, true);
+        let openai_gpt_tokenizer: OpenAiGptTokenizer =
+            OpenAiGptTokenizer::from_existing_vocab_and_merges(vocab, merges, true);
         let skip_special_tokens = false;
         let clean_up_tokenization_spaces = false;
-        let test_tuples = [(vec![4, 8, 9], "the earth")];
+        let test_tuples = [(vec![8, 10, 9], "the earth")];
         let source_ids: Vec<Vec<i64>> = test_tuples.iter().map(|v| v.0.clone()).collect_vec();
         let expected_results: Vec<&str> = test_tuples.iter().map(|v| v.1.clone()).collect_vec();
 
         //        When & Then
         for (source_ids, expected_result) in test_tuples.iter() {
             assert_eq!(
-                gpt2_tokenizer.decode(
+                openai_gpt_tokenizer.decode(
                     source_ids.clone(),
                     skip_special_tokens,
                     clean_up_tokenization_spaces
@@ -349,7 +315,7 @@ mod tests {
         }
         assert_eq!(
             Tokenizer::decode_list(
-                &gpt2_tokenizer,
+                &openai_gpt_tokenizer,
                 source_ids.clone(),
                 skip_special_tokens,
                 clean_up_tokenization_spaces
