+++ main/src/vocab/bpe_vocab.rs
@@ -10,7 +10,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::preprocessing::error::TokenizerError;
+use crate::error::TokenizerError;
 use std::collections::HashMap;
 use std::fs::File;
 use std::io::{BufRead, BufReader};
