+++ main/src/vocab/gpt2_vocab.rs
@@ -11,8 +11,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::preprocessing::error::TokenizerError;
-use crate::preprocessing::vocab::base_vocab::{swap_key_values, Vocab};
+use crate::error::TokenizerError;
+use crate::vocab::base_vocab::{swap_key_values, Vocab};
 use std::collections::HashMap;
 use std::fs::File;
 use std::io::BufReader;
