+++ main/src/vocab/xlnet_vocab.rs
@@ -1,5 +1,4 @@
-// Copyright 2018 Google AI, Google Brain and Carnegie Mellon University Authors.
-// Copyright 2018-2020 The HuggingFace Inc. team.
+// Copyright 2018 Google AI, Google Brain and Carnegie Mellon University Authors and the HuggingFace Inc. team.
 // Copyright 2019-2020 Guillaume Becquin
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -11,16 +10,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::preprocessing::error::TokenizerError;
-use crate::preprocessing::vocab::base_vocab::swap_key_values;
-use crate::preprocessing::vocab::sentencepiece_proto::sentencepiece_model::ModelProto;
-use crate::Vocab;
+use crate::error::TokenizerError;
+use crate::vocab::base_vocab::swap_key_values;
+use crate::vocab::sentencepiece_proto::sentencepiece_model::ModelProto;
+use crate::vocab::Vocab;
 use protobuf::parse_from_bytes;
 use std::collections::HashMap;
 use std::fs::File;
 use std::io::Read;
 
-pub struct XLMRobertaVocab {
+pub struct XLNetVocab {
     pub values: HashMap<String, i64>,
     pub indices: HashMap<i64, String>,
     pub unknown_value: &'static str,
@@ -28,7 +27,7 @@ pub struct XLMRobertaVocab {
     pub special_indices: HashMap<i64, String>,
 }
 
-impl XLMRobertaVocab {
+impl XLNetVocab {
     pub fn bos_value() -> &'static str {
         "<s>"
     }
@@ -36,10 +35,10 @@ impl XLMRobertaVocab {
         "</s>"
     }
     pub fn sep_value() -> &'static str {
-        "</s>"
+        "<sep>"
     }
     pub fn cls_value() -> &'static str {
-        "<s>"
+        "<cls>"
     }
     pub fn mask_value() -> &'static str {
         "<mask>"
@@ -47,9 +46,15 @@ impl XLMRobertaVocab {
     pub fn pad_value() -> &'static str {
         "<pad>"
     }
+    pub fn eop_value() -> &'static str {
+        "<eop>"
+    }
+    pub fn eod_value() -> &'static str {
+        "<eod>"
+    }
 }
 
-impl Vocab for XLMRobertaVocab {
+impl Vocab for XLNetVocab {
     fn unknown_value() -> &'static str {
         "<unk>"
     }
@@ -74,7 +79,7 @@ impl Vocab for XLMRobertaVocab {
         &self.special_indices
     }
 
-    fn from_file(path: &str) -> Result<XLMRobertaVocab, TokenizerError> {
+    fn from_file(path: &str) -> Result<XLNetVocab, TokenizerError> {
         let mut f = File::open(path).map_err(|e| {
             TokenizerError::FileNotFound(format!("{} vocabulary file not found :{}", path, e))
         })?;
@@ -90,45 +95,44 @@ impl Vocab for XLMRobertaVocab {
                 return Err(TokenizerError::VocabularyParsingError(e.to_string()));
             }
         };
+
         let mut values = HashMap::new();
-        values.insert(XLMRobertaVocab::cls_value().to_owned(), values.len() as i64);
-        values.insert(XLMRobertaVocab::pad_value().to_owned(), values.len() as i64);
-        values.insert(XLMRobertaVocab::eos_value().to_owned(), values.len() as i64);
-        values.insert(
-            XLMRobertaVocab::unknown_value().to_owned(),
-            values.len() as i64,
-        );
-        for piece in proto.get_pieces().iter().skip(3) {
-            values.insert(piece.get_piece().to_owned(), values.len() as i64);
+        for (idx, piece) in proto.get_pieces().iter().enumerate() {
+            values.insert(piece.get_piece().to_owned(), idx as i64);
         }
-        values.insert(
-            XLMRobertaVocab::mask_value().to_owned(),
-            values.len() as i64,
-        );
 
         let mut special_values = HashMap::new();
-        let unknown_value = XLMRobertaVocab::unknown_value();
-        XLMRobertaVocab::_register_as_special_value(unknown_value, &values, &mut special_values)?;
+        let unknown_value = XLNetVocab::unknown_value();
+        XLNetVocab::_register_as_special_value(unknown_value, &values, &mut special_values)?;
 
-        let bos_value = XLMRobertaVocab::bos_value();
-        XLMRobertaVocab::_register_as_special_value(bos_value, &values, &mut special_values)?;
+        let bos_value = XLNetVocab::bos_value();
+        XLNetVocab::_register_as_special_value(bos_value, &values, &mut special_values)?;
 
-        let eos_value = XLMRobertaVocab::eos_value();
-        XLMRobertaVocab::_register_as_special_value(eos_value, &values, &mut special_values)?;
+        let eos_value = XLNetVocab::eos_value();
+        XLNetVocab::_register_as_special_value(eos_value, &values, &mut special_values)?;
 
-        let cls_value = XLMRobertaVocab::cls_value();
-        XLMRobertaVocab::_register_as_special_value(cls_value, &values, &mut special_values)?;
+        let cls_value = XLNetVocab::cls_value();
+        XLNetVocab::_register_as_special_value(cls_value, &values, &mut special_values)?;
 
-        let mask_value = XLMRobertaVocab::mask_value();
-        XLMRobertaVocab::_register_as_special_value(mask_value, &values, &mut special_values)?;
+        let mask_value = XLNetVocab::mask_value();
+        XLNetVocab::_register_as_special_value(mask_value, &values, &mut special_values)?;
 
-        let pad_value = XLMRobertaVocab::pad_value();
-        XLMRobertaVocab::_register_as_special_value(pad_value, &values, &mut special_values)?;
+        let pad_value = XLNetVocab::pad_value();
+        XLNetVocab::_register_as_special_value(pad_value, &values, &mut special_values)?;
+
+        let sep_value = XLNetVocab::sep_value();
+        XLNetVocab::_register_as_special_value(sep_value, &values, &mut special_values)?;
+
+        let eop_value = XLNetVocab::eop_value();
+        XLNetVocab::_register_as_special_value(eop_value, &values, &mut special_values)?;
+
+        let eod_value = XLNetVocab::eod_value();
+        XLNetVocab::_register_as_special_value(eod_value, &values, &mut special_values)?;
 
         let indices = swap_key_values(&values);
         let special_indices = swap_key_values(&special_values);
 
-        Ok(XLMRobertaVocab {
+        Ok(XLNetVocab {
             values,
             indices,
             unknown_value,
