+++ main/tests/test_albert_uncased.rs
@@ -1,9 +1,9 @@
-use rust_tokenizers::preprocessing::tokenizer::base_tokenizer::Offset;
-use rust_tokenizers::{TokenizedInput, Tokenizer, TruncationStrategy};
-
 mod test_utils;
 
-use rust_tokenizers::preprocessing::tokenizer::albert_tokenizer::AlbertTokenizer;
+use rust_tokenizers::tokenizer::{
+    AlbertTokenizer, MultiThreadedTokenizer, Tokenizer, TruncationStrategy,
+};
+use rust_tokenizers::{Offset, TokenizedInput};
 use test_utils::download_file_to_cache;
 
 #[test]
@@ -206,7 +206,8 @@ fn test_albert_tokenization() -> anyhow::Result<()> {
     ]
     .to_vec();
 
-    let output = albert_tokenizer.encode_list(
+    let output = MultiThreadedTokenizer::encode_list(
+        &albert_tokenizer,
         original_strings.to_vec(),
         128,
         &TruncationStrategy::LongestFirst,
