+++ main/tests/test_openai_gpt_uncased.rs
@@ -1,8 +1,7 @@
-use rust_tokenizers::preprocessing::tokenizer::base_tokenizer::Offset;
-use rust_tokenizers::{OpenAiGptTokenizer, TokenizedInput, Tokenizer, TruncationStrategy};
-
 mod test_utils;
 
+use rust_tokenizers::tokenizer::{OpenAiGptTokenizer, Tokenizer, TruncationStrategy};
+use rust_tokenizers::{Offset, TokenizedInput};
 use test_utils::download_file_to_cache;
 
 #[test]
