+++ main/tests/test_t5_cased.rs
@@ -1,10 +1,10 @@
 extern crate anyhow;
 
 mod test_utils;
-
-use rust_tokenizers::preprocessing::tokenizer::base_tokenizer::Offset;
-use rust_tokenizers::preprocessing::tokenizer::t5_tokenizer::T5Tokenizer;
-use rust_tokenizers::{TokenizedInput, Tokenizer, TruncationStrategy};
+use rust_tokenizers::tokenizer::{
+    MultiThreadedTokenizer, T5Tokenizer, Tokenizer, TruncationStrategy,
+};
+use rust_tokenizers::{Offset, TokenizedInput};
 use test_utils::download_file_to_cache;
 
 #[test]
@@ -225,7 +225,8 @@ fn test_t5_tokenization() -> anyhow::Result<()> {
     ]
     .to_vec();
 
-    let output = t5_tokenizer.encode_list(
+    let output = MultiThreadedTokenizer::encode_list(
+        &t5_tokenizer,
         original_strings.to_vec(),
         128,
         &TruncationStrategy::LongestFirst,
